Enum datafusion_expr::BuiltinScalarFunction

pub enum BuiltinScalarFunction {

    Abs,
    Acos,
    Asin,
    Atan,
    Atan2,
    Ceil,
    Coalesce,
    Cos,
    Digest,
    Exp,
    Floor,
    Ln,
    Log,
    Log10,
    Log2,
    Power,
    Round,
    Signum,
    Sin,
    Sqrt,
    Tan,
    Trunc,
    MakeArray,
    Ascii,
    BitLength,
    Btrim,
    CharacterLength,
    Chr,
    Concat,
    ConcatWithSeparator,
    DatePart,
    DateTrunc,
    DateBin,
    InitCap,
    Left,
    Lpad,
    Lower,
    Ltrim,
    MD5,
    NullIf,
    OctetLength,
    Random,
    RegexpReplace,
    Repeat,
    Replace,
    Reverse,
    Right,
    Rpad,
    Rtrim,
    SHA224,
    SHA256,
    SHA384,
    SHA512,
    SplitPart,
    StartsWith,
    Strpos,
    Substr,
    ToHex,
    ToTimestamp,
    ToTimestampMillis,
    ToTimestampMicros,
    ToTimestampSeconds,
    FromUnixtime,
    Now,
    CurrentDate,
    CurrentTime,
    Translate,
    Trim,
    Upper,
    Uuid,
    RegexpMatch,
    Struct,
    ArrowTypeof,

}

Enum of all built-in scalar functions

Variants

Abs

abs

Acos

acos

Asin

asin

Atan

atan

Atan2

atan2

Ceil

ceil

Coalesce

coalesce

Cos

cos

Digest

Digest

Exp

exp

Floor

floor

Ln

ln, Natural logarithm

Log

log, same as log10

Log10

log10

Log2

log2

Power

power

Round

round

Signum

signum

Sin

sin

Sqrt

sqrt

Tan

tan

Trunc

trunc

MakeArray

construct an array from columns

Ascii

ascii

BitLength

bit_length

Btrim

btrim

CharacterLength

character_length

Chr

chr

Concat

concat

ConcatWithSeparator

concat_ws

DatePart

date_part

DateTrunc

date_trunc

DateBin

date_bin

InitCap

initcap

Left

left

Lpad

lpad

Lower

lower

Ltrim

ltrim

MD5

md5

NullIf

nullif

OctetLength

octet_length

Random

random

RegexpReplace

regexp_replace

Repeat

repeat

Replace

replace

Reverse

reverse

Right

right

Rpad

rpad

Rtrim

rtrim

SHA224

sha224

SHA256

sha256

SHA384

sha384

SHA512

Sha512

SplitPart

split_part

StartsWith

starts_with

Strpos

strpos

Substr

substr

ToHex

to_hex

ToTimestamp

to_timestamp

ToTimestampMillis

to_timestamp_millis

ToTimestampMicros

to_timestamp_micros

ToTimestampSeconds

to_timestamp_seconds

FromUnixtime

from_unixtime

Now

now

CurrentDate

current_date

CurrentTime

current_time

Translate

translate

Trim

trim

Upper

upper

Uuid

uuid

RegexpMatch

regexp_match

Struct

struct

ArrowTypeof

arrow_typeof

Implementations

impl BuiltinScalarFunction

pub fn supports_zero_argument(&self) -> bool

an allowlist of functions to take zero arguments, so that they will get special treatment while executing.

pub fn volatility(&self) -> Volatility

Returns the Volatility of the builtin function.