Module datafusion_expr::expr_rewriter

Expression rewriter

Enums

RewriteRecursion

Controls how the ExprRewriter recursion should proceed.

Traits

ExprRewritable

A trait for marking types that are rewritable by ExprRewriter

ExprRewriter

Trait for potentially recursively rewriting an Expr expression tree. When passed to Expr::rewrite, ExpressionVisitor::mutate is invoked recursively on all nodes of an expression tree.

Functions

coerce_plan_expr_for_schema

Returns plan with expressions coerced to types compatible with schema types

normalize_col

Recursively call Column::normalize_with_schemas on all Column expressions in the expr expression tree.

normalize_col_with_schemas

Recursively call Column::normalize_with_schemas on all Column expressions in the expr expression tree.

normalize_cols

Recursively normalize all Column expressions in a list of expression trees

replace_col

Recursively replace all Column expressions in a given expression tree with Column expressions provided by the hash map argument.

rewrite_expr

Recursively rewrite an Expr via a function.

rewrite_sort_cols_by_aggs

Rewrite sort on aggregate expressions to sort on the column of aggregate output For example, max(x) is written to col("MAX(x)")

unnormalize_col

Recursively ‘unnormalize’ (remove all qualifiers) from an expression tree.

unnormalize_cols

Recursively un-normalize all Column expressions in a list of expression trees