Module datafusion_expr::logical_plan

Re-exports

pub use builder::table_scan;

pub use builder::LogicalPlanBuilder;

pub use display::display_schema;

Modules

builder

This module provides a builder for creating LogicalPlans

display

This module provides logic for displaying LogicalPlans in various styles

Structs

Aggregate

Aggregates its input based on a set of grouping and aggregate expressions (e.g. SUM).

Analyze

Runs the actual plan, and then prints the physical plan with with execution metrics.

Boolean

CreateCatalog

Creates a catalog (aka “Database”).

CreateCatalogSchema

Creates a schema.

CreateExternalTable

Creates an external table.

CreateMemoryTable

Creates an in memory table.

CreateView

Creates a view.

CrossJoin

Apply Cross Join to two logical plans

DescribeTable

Describe the schema of table

Distinct

Removes duplicate rows from the input

DmlStatement

The operator that modifies the content of a database (adapted from substrait WriteRel)

DropTable

Drops a table.

DropView

Drops a view.

EmptyRelation

Produces no rows: An empty relation with an empty schema

Explain

Produces a relation with string representations of various parts of the plan

Extension

Extension operator defined outside of DataFusion

Filter

Filters rows from its input that do not match an expression (essentially a WHERE clause with a predicate expression).

Join

Join two logical plans on one or more join columns

Limit

Produces the first n tuples from its input and discards the rest.

Prepare

Prepare a statement but do not execute it. Prepare statements can have 0 or more Expr::Placeholder expressions that are filled in during execution

Projection

Evaluates an arbitrary list of expressions (essentially a SELECT with an expression list) on its input.

Repartition

Repartition the plan based on a partitioning scheme.

SetVariable

Set a Variable’s value – value in ConfigOptions

Sort

Sorts its input according to a list of sort expressions.

StringifiedPlan

Represents some sort of execution plan, in String form

Subquery

Subquery

SubqueryAlias

Aliased subquery

TableScan

Produces rows from a table provider by reference or from the context

Union

Union multiple inputs

Unnest

Unnest a column that contains a nested list type.

Values

Values expression. See Postgres VALUES documentation for more details.

Window

Window its input based on a set of window spec and window function (e.g. SUM or RANK)

Enums

JoinConstraint

Join constraint

JoinType

Join type

LogicalPlan

A LogicalPlan represents the different types of relational operators (such as Projection, Filter, etc) and can be created by the SQL query planner and the DataFrame API.

Partitioning

Logical partitioning schemes supported by the repartition operator.

PlanType

Represents which type of plan, when storing multiple for use in EXPLAIN plans

Predicate

WriteOp

Traits

PlanVisitor

Trait that implements the Visitor pattern for a depth first walk of LogicalPlan nodes. pre_visit is called before any children are visited, and then post_visit is called after all children have been visited. To use, define a struct that implements this trait and then invoke LogicalPlan::accept.

ToStringifiedPlan

Trait for something that can be formatted as a stringified plan

UserDefinedLogicalNode

This defines the interface for LogicalPlan nodes that can be used to extend DataFusion with custom relational operators.