Module datafusion_expr::type_coercion::binary

Coercion rules for matching argument types for binary operators

Functions

binary_operator_data_type

Returns the return type of a binary operator or an error when the binary operator cannot perform the computation between the argument’s types, even after type coercion.

coerce_types

Coercion rules for all binary operators. Returns the output type of applying op to an argument of lhs_type and rhs_type.

comparison_coercion

Returns the output type of applying comparison operations such as eq, not eq, lt, lteq, gt, and gteq to arguments of lhs_type and rhs_type.

like_coercion

coercion rules for like operations. This is a union of string coercion rules and dictionary coercion rules

regex_coercion

coercion rules for regular expression comparison operations. This is a union of string coercion rules and dictionary coercion rules