Module datafusion_expr::utils

Expression utilities

Constants

COUNT_STAR_EXPANSION

The value to which COUNT(*) is expanded to in COUNT(<constant>) expressions

Functions

can_hash

can this data type be used in hash join equal conditions?? data types here come from function ‘equal_rows’, if more data types are supported in equal_rows(hash join), add those data types here to generate join logical plan.

check_all_column_from_schema

Check whether all columns are from the schema.

columnize_expr

Convert an expression into Column expression if it’s already provided as input plan.

compare_sort_expr

Compare the sort expr as PostgreSQL’s common_prefix_cmp(): https://github.com/postgres/postgres/blob/master/src/backend/optimizer/plan/planner.c

ensure_any_column_reference_is_unambiguous

Ensure any column reference of the expression is unambiguous. Assume we have two schema: schema1: a, b ,c schema2: a, d, e

expand_qualified_wildcard

Resolves an Expr::Wildcard to a collection of qualified Expr::Column’s.

expand_wildcard

Resolves an Expr::Wildcard to a collection of Expr::Column’s.

expr_as_column_expr

Convert any Expr to an Expr::Column.

expr_to_columns

Recursively walk an expression tree, collecting the unique set of columns referenced in the expression

exprlist_to_columns

Recursively walk a list of expression trees, collecting the unique set of columns referenced in the expression

exprlist_to_fields

Create field meta-data from an expression, for use in a result set schema

find_aggregate_exprs

Collect all deeply nested Expr::AggregateFunction and Expr::AggregateUDF. They are returned in order of occurrence (depth first), with duplicates omitted.

find_column_exprs

Collect all deeply nested Expr::Column’s. They are returned in order of appearance (depth first), and may contain duplicates.

find_sort_exprs

Collect all deeply nested Expr::Sort. They are returned in order of occurrence (depth first), with duplicates omitted.

find_valid_equijoin_key_pair

Give two sides of the equijoin predicate, return a valid join key pair. If there is no valid join key pair, return None.

find_window_exprs

Collect all deeply nested Expr::WindowFunction. They are returned in order of occurrence (depth first), with duplicates omitted.

from_plan

Returns a new logical plan based on the original one with inputs and expressions replaced.

generate_sort_key

Generate a sort key for a given window expr’s partition_by and order_bu expr

group_window_expr_by_sort_keys

group a slice of window expression expr by their order by expressions

grouping_set_expr_count

Count the number of distinct exprs in a list of group by expressions. If the first element is a GroupingSet expression then it must be the only expr.

grouping_set_to_exprlist

Find all distinct exprs in a list of group by expressions. If the first element is a GroupingSet expression then it must be the only expr.